blocks/course_template/ Fix 'course_template' plugin from breaking core tests
The 'blocks_course_template' plugin was breaking core tests due to one
of the columns in the 'block_course_template' table using the reserved
word 'file' for its name. This commit changes that reserved word to
'filename'.

The commit also includes an upgrade script to be run next time the
upgrade.php cli script is run. This will make the name change to the
database.